simplify Vuex and vue-router of demo application
